working on filterToken, filter somehow working?

+++ RealEstate/Models/DatabaseModels/Offer.cs
@@ -12,7 +12,7 @@ namespace RealEstate.Models.DatabaseModels
         public int Id { get; set; }
         public int? IdBroker { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+        public ulong Price { get; set; }
         public string Location { get; set; }
         public string Description { get; set; }
         [MaxLength(100)]
