+++ RealEstate/Models/ViewModels/OfferList.cs
@@ -7,7 +7,7 @@ namespace RealEstate.Models.ViewModels
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+        public ulong Price { get; set; }
         public string Location { get; set; }
         public string? Broker { get; set; }
         public string Category { get; set; } // f - flats, l - luxury,  h - houses, c - cottages
